requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
sqlalchemy==2.0.45
asyncpg==0.31.0
greenlet==3.1.1